split sub.rs into _iterable and _self
