+++ src/impls.rs
@@ -1,5 +1,6 @@
 mod add;
 mod index;
 mod intersection;
-mod sub;
+mod sub_iterable;
+mod sub_self;
 mod union;
